Refactor DefaultPipelineLauncher
