+++ it/common/src/test/java/org/apache/beam/it/common/dataflow/IOStressTestBase.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.it.common;
+package org.apache.beam.it.common.dataflow;
 
 import java.io.Serializable;
 import java.time.Duration;
