+++ it/google-cloud-platform/src/test/java/org/apache/beam/it/gcp/bigquery/BigQueryIOLT.java
@@ -45,7 +45,7 @@ import org.apache.beam.it.common.PipelineOperator;
 import org.apache.beam.it.common.TestProperties;
 import org.apache.beam.it.common.bigquery.BigQueryResourceManager;
 import org.apache.beam.it.common.utils.ResourceManagerUtils;
-import org.apache.beam.it.common.IOLoadTestBase;
+import org.apache.beam.it.common.dataflow.IOLoadTestBase;
 import org.apache.beam.sdk.io.Read;
 import org.apache.beam.sdk.io.gcp.bigquery.AvroWriteRequest;
 import org.apache.beam.sdk.io.gcp.bigquery.BigQueryIO;
