+++ it/iceberg/src/test/java/org/apache/beam/it/iceberg/IcebergIOLT.java
@@ -1,6 +1,6 @@
 package org.apache.beam.it.iceberg;
 
-import org.apache.beam.it.common.IOLoadTestBase;
+import org.apache.beam.it.common.dataflow.IOLoadTestBase;
 
 public class IcebergIOLT extends IOLoadTestBase {
