+++ it/kafka/build.gradle
@@ -42,8 +42,7 @@ dependencies {
     permitUsedUndeclared library.java.guava
     testImplementation library.java.mockito_core
     testRuntimeOnly library.java.slf4j_simple
-    // TODO: Remove the below dependencies after KafkaIOLT has been moved to the right directory.
-    testImplementation project(path: ":it:google-cloud-platform")
+    testImplementation project(path: ":it:common", configuration: "testRuntimeMigration")
     testImplementation project(path: ":sdks:java:io:kafka")
     testImplementation project(path: ":sdks:java:io:synthetic")
     testImplementation project(path: ":runners:direct-java")
